requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
Faker==13.12.0
graphql-core==3.2.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyHamcrest==2.0.3
pyparsing==3.0.9
pytest==6.2.4
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.0
sgqlc==13.0
six==1.16.0
toml==0.10.2
urllib3==1.26.9

openpyxl~=3.0.10